requirements.txt
flask==3.0.0
beautifulsoup4==4.12.2
selenium==4.16.0
lxml
html5lib
python-dotenv==1.0.0
google-generativeai==0.3.1